섬의개수 시도....... 그러나 실패!

+++ baekjoonRe/1032 명령 프롬프트.cpp
@@ -3,7 +3,7 @@
 
 using namespace std;
 
-int n,len,i,j;
+unsigned long n,len,i,j;
 char arr[55][55];
 bool flag;
